findAll 테스트 수정
